+++ ui/src/records/recordUpsertModal.js
@@ -421,7 +421,11 @@ function recordUpsertModal(collection, rawRecord, modalSettings) {
                         inert: () => data.isLoading || data.isSaving,
                         onmount: (el) => {
                             el._quickSaveHandler = (e) => {
-                                if ((e.ctrlKey || e.metaKey) && e.code == "KeyS") {
+                                if (
+                                    (e.ctrlKey || e.metaKey)
+                                    && e.code == "KeyS"
+                                    && app.modals.getTop() === el?.closest(".modal")
+                                ) {
                                     e.preventDefault();
                                     save(false);
                                 }
@@ -430,7 +434,7 @@ function recordUpsertModal(collection, rawRecord, modalSettings) {
                         },
                         onunmount: (el) => {
                             if (el?._quickSaveHandler) {
-                                window.removeEventListener("keydown", el?._quickSaveHandler);
+                                window.removeEventListener("keydown", el._quickSaveHandler);
                             }
                         },
                     },